feat: disable possible data operations during running mutation with data lists on admin dashboard
